refactor: create api package
