+++ src/main/kotlin/com/jonaswanke/unicorn/api/Git.kt
@@ -1,4 +1,4 @@
-package com.jonaswanke.unicorn.script
+package com.jonaswanke.unicorn.api
 
 import com.jonaswanke.unicorn.commands.RunContext
 import com.jonaswanke.unicorn.utils.code
