+++ src/main/kotlin/com/jonaswanke/unicorn/commands/Login.kt
@@ -3,7 +3,7 @@ package com.jonaswanke.unicorn.commands
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.optional
 import com.github.ajalt.clikt.parameters.options.option
-import com.jonaswanke.unicorn.script.GitHub
+import com.jonaswanke.unicorn.api.GitHub
 
 open class Login : BaseCommand() {
     private val username by argument("username").optional()
